Nimble: Fix crash when attempting to do deinit
If esp timer is created, but not started, and if we call Nimble deinit , then during deinit, the existing code checks for the
timer state and returns with a status other than ESP_OK. However, existing ESP_ERROR_CHECK macro treats this as failure and causes crash.

Modified the code to use ESP_ERROR_CHECK_WITHOUT_ABORT instead.